refactor: PlayerGameResult에 UnaryOperator 적용 및 패키지 분리
