+++ src/test/java/domain/result/PlayerGameResultTest.java
@@ -1,8 +1,8 @@
-package domain.participant;
+package domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.participant.PlayerGameResult;
+import domain.result.PlayerGameResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
